Helper functions for testing moved into package dae.testing
WIP #230
