+++ dae/dae/genomic_resources/tests/test_gene_models_resource.py
@@ -5,7 +5,7 @@ import pytest
 from dae.genomic_resources.gene_models import \
     build_gene_models_from_resource, GeneModels
 from dae.genomic_resources.testing import build_test_resource
-from dae.genomic_resources.testing import convert_to_tab_separated
+from dae.testing import convert_to_tab_separated
 from dae.genomic_resources.fsspec_protocol import build_fsspec_protocol
 from dae.genomic_resources.repository import GenomicResourceProtocolRepo, \
     GenomicResource
