+++ dae/dae/genomic_resources/tests/test_score_resource.py
@@ -1,7 +1,7 @@
 # pylint: disable=W0621,C0114,C0116,W0212,W0613
 
 from dae.genomic_resources.testing import build_testing_repository
-from dae.genomic_resources.testing import convert_to_tab_separated
+from dae.testing import convert_to_tab_separated
 
 
 def test_pos_gr(tmp_path):
