+++ dae/dae/impala_storage/schema1/tests/test_impala_variants_person_set_queries.py
@@ -7,7 +7,7 @@ import toml
 import pytest
 
 from dae.impala_storage.schema1.impala_variants import ImpalaVariants
-from dae.genomic_resources.testing import convert_to_tab_separated
+from dae.testing import convert_to_tab_separated
 from dae.configuration.gpf_config_parser import GPFConfigParser
 from dae.configuration.schemas.person_sets import person_set_collections_schema
 from dae.pedigrees.loader import FamiliesLoader
