+++ dae/dae/pedigrees/testing.py
@@ -3,7 +3,7 @@
 import io
 import textwrap
 
-from dae.genomic_resources.testing import convert_to_tab_separated
+from dae.testing import convert_to_tab_separated
 
 from dae.pedigrees.family import FamiliesData, Family
 from dae.pedigrees.loader import FamiliesLoader
