+++ dae/dae/testing/__init__.py
@@ -1,0 +1,10 @@
+from .helpers import convert_to_tab_separated, setup_directories, \
+    setup_pedigree, setup_vcf, setup_denovo, setup_dae_transmitted, \
+    setup_genome, setup_gene_models, setup_empty_gene_models
+
+
+__all__ = [
+    "convert_to_tab_separated", "setup_directories",
+    "setup_pedigree", "setup_vcf", "setup_denovo", "setup_dae_transmitted",
+    "setup_genome", "setup_gene_models", "setup_empty_gene_models",
+]
