+++ dae/dae/variants_loaders/cnv/tests/test_configure_cnv_loader.py
@@ -3,7 +3,7 @@ import io
 import textwrap
 import pytest
 
-from dae.genomic_resources.testing import convert_to_tab_separated
+from dae.testing import convert_to_tab_separated
 from dae.pedigrees.loader import FamiliesLoader
 
 from dae.variants_loaders.cnv.flexible_cnv_loader import \
