+++ dae/dae/variants_loaders/cnv/tests/test_flexible_cnv_loader.py
@@ -4,7 +4,7 @@ import textwrap
 import pytest
 
 from dae.utils.variant_utils import mat2str
-from dae.genomic_resources.testing import convert_to_tab_separated
+from dae.testing import convert_to_tab_separated
 
 from dae.variants.core import Allele
 from dae.variants_loaders.cnv.flexible_cnv_loader import flexible_cnv_loader
