+++ dae/tests/integration/query_variants/test_best_state.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from tests.foobar_import import setup_vcf, setup_pedigree
+from dae.testing import setup_vcf, setup_pedigree
 
 from dae.utils.variant_utils import mat2str
 from dae.utils.regions import Region
